Re-added game launch behaviour
